requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-admin-thumbnails==0.2.5
django-appconf==1.0.4
django-compressor==2.4.1
flake8==3.9.2
flake8-isort==4.0.0
idna==2.10
isort==5.8.0
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.1
testfixtures==6.17.1
toml==0.10.2
urllib3==1.26.5
whitenoise==5.2.0